requirements.txt
# ULTRA-LOW LATENCY OPTIMIZED DEPENDENCIES
# Pin versions for stability and performance

# LiveKit and Agents - Core voice processing
livekit>=0.16.3
livekit-agents>=0.9.0
livekit-plugins-openai>=0.7.0
livekit-plugins-silero>=0.6.0

# Twilio - Phone integration
twilio>=9.2.0

# Web Framework - PERFORMANCE OPTIMIZED
fastapi>=0.109.0
uvicorn[standard]>=0.27.0  # [standard] includes performance optimizations
python-multipart>=0.0.6

# PERFORMANCE: Ultra-fast event loop (critical for low latency)
# NOTE: uvloop only works on Linux/Mac - Windows will use standard asyncio
uvloop>=0.19.0; sys_platform != 'win32'  # Faster than asyncio (Unix only)

# Environment variables
python-dotenv>=1.0.0

# Async HTTP - PERFORMANCE OPTIMIZED
aiohttp>=3.10.0
httpx>=0.27.0  # Fast HTTP client

# Audio processing - OPTIMIZED
numpy>=1.24.0,<2.0.0  # Pin to 1.x for performance
audioop-lts>=0.2.1; python_version >= '3.13'  # Only needed for Python 3.13+

# Logging - Minimal overhead
colorlog>=6.8.0

# Optional: Performance monitoring
psutil>=5.9.0  # System resource monitoring
orjson>=3.9.0  # Faster JSON encoding/decoding

